Web - Facebook Sign In

+++ template/src/app/auth-providers/Facebook.tsx
@@ -1,6 +1,6 @@
-import auth from '@react-native-firebase/auth';
+import auth, {facebookWebSignIn} from '@react-native-firebase/auth';
 import {useContext, useState} from 'react';
-import {Alert} from 'react-native';
+import {Alert, Platform} from 'react-native';
 import {AccessToken, LoginManager} from 'react-native-fbsdk-next';
 import {UserContext} from '../App';
 import ProviderButton from '../components/AuthProviderButton';
@@ -20,43 +20,46 @@ function Facebook(): JSX.Element | null {
   async function handleFacebook() {
     if (!loading) {
       setLoading(true);
-
-      try {
-        if (variant === 'UNLINK' && user) {
-          await user.unlink(PROVIDER_ID);
-          await user.reload();
-        } else {
-          const {isCancelled} = await LoginManager.logInWithPermissions([
-            'public_profile',
-            'email',
-          ]);
-
-          if (isCancelled) {
-            Alert.alert('Facebook Auth Canceled');
+      if (Platform.OS === 'web') {
+        await facebookWebSignIn();
+      } else {
+        try {
+          if (variant === 'UNLINK' && user) {
+            await user.unlink(PROVIDER_ID);
+            await user.reload();
           } else {
-            const result = await AccessToken.getCurrentAccessToken();
-            if (!result) {
-              throw new Error(
-                'No Access Token was returned from Facebook SDK.',
-              );
-            }
+            const {isCancelled} = await LoginManager.logInWithPermissions([
+              'public_profile',
+              'email',
+            ]);
 
-            const {accessToken} = result;
+            if (isCancelled) {
+              Alert.alert('Facebook Auth Canceled');
+            } else {
+              const result = await AccessToken.getCurrentAccessToken();
+              if (!result) {
+                throw new Error(
+                  'No Access Token was returned from Facebook SDK.',
+                );
+              }
 
-            const credential =
-              auth.FacebookAuthProvider.credential(accessToken);
+              const {accessToken} = result;
 
-            if (variant === 'LINK' && user) {
-              await user.linkWithCredential(credential);
-              await user.reload();
-            } else if (variant === 'SIGN_IN') {
-              await auth().signInWithCredential(credential);
+              const credential =
+                auth.FacebookAuthProvider.credential(accessToken);
+
+              if (variant === 'LINK' && user) {
+                await user.linkWithCredential(credential);
+                await user.reload();
+              } else if (variant === 'SIGN_IN') {
+                await auth().signInWithCredential(credential);
+              }
             }
           }
+        } catch (error) {
+          setLoading(false);
+          Alert.alert('Facebook Auth Error', (error as Error).message);
         }
-      } catch (error) {
-        setLoading(false);
-        Alert.alert('Facebook Auth Error', (error as Error).message);
       }
     }
   }
